refactor: merge beta and refactor source code
