+++ example/App.d.ts
@@ -4,5 +4,6 @@
  *
  * @format
  */
+/// <reference types="react" />
 declare function App(): JSX.Element;
 export default App;
