+++ src/types/constant.d.ts
@@ -1,0 +1,21 @@
+import { ImageSourcePropType } from 'react-native';
+declare const SCREENSHOT_EVT = "onScreenShotCaptured";
+declare const SCREEN_RECORDING_EVT = "onScreenRecordingCaptured";
+declare const BLACK_COLOR = "#000000";
+declare const REGEX: RegExp;
+declare const IMAGE_REGEX: RegExp;
+declare const RADIUS_DEFAULT = 15;
+declare const TIME_DELAYED = 1000;
+declare const Alignment: {
+    topLeft: number;
+    topCenter: number;
+    topRight: number;
+    centerLeft: number;
+    center: number;
+    centerRight: number;
+    bottomLeft: number;
+    bottomCenter: number;
+    bottomRight: number;
+};
+declare const resolveAssetSource: (defaultSource: ImageSourcePropType | any) => any;
+export { SCREENSHOT_EVT, SCREEN_RECORDING_EVT, BLACK_COLOR, REGEX, IMAGE_REGEX, Alignment, RADIUS_DEFAULT, TIME_DELAYED, resolveAssetSource, };
