+++ tsconfig.build.json
@@ -1,5 +1,5 @@
 
 {
   "extends": "./tsconfig",
-  "exclude": ["example"]
+  "exclude": ["example", "node_modules"]
 }
