+++ tsconfig.json
@@ -2,14 +2,16 @@
   "compilerOptions": {
     "baseUrl": "./",
     "paths": {
-      "react-native-screenguard": ["index"]
+      "react-native-screenguard": ["./src/commonjs/index"]
     },
-    "declaration": true,
+    "outDir": "./src/commonjs",      
+    "declaration": true,             
+    "declarationDir": "./src/types", 
+    "rootDir": "./src/main",         
+    "esModuleInterop": true,         
+    "forceConsistentCasingInFileNames": true,
     "allowUnreachableCode": false,
     "allowUnusedLabels": false,
-    "esModuleInterop": true,
-    "declaration": true,
-    "forceConsistentCasingInFileNames": true,
     "jsx": "react",
     "lib": ["esnext"],
     "module": "esnext",
@@ -24,6 +26,13 @@
     "resolveJsonModule": true,
     "skipLibCheck": true,
     "strict": true,
-    "target": "esnext"
-  }
+    "target": "esnext",
+  },
+  "include": [
+    "src/main/**/*.ts"
+  ],
+  "exclude": [
+    "node_modules",   
+    "example"         
+  ]
 }
